crog-ai-backend/pyproject.toml
[project]
name = "crog-ai-backend"
version = "0.1.0"
description = "CROG-AI command center: Market Club strangler + Dochia signal engine"
requires-python = ">=3.12"

dependencies = [
    "sqlalchemy>=2.0.30",
    "alembic>=1.13.0",
    "psycopg[binary]>=3.2.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "numpy>=2.0.0",
    "polars>=1.5.0",
    "scipy>=1.14.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "imap-tools>=1.7.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "RET"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_mode = "auto"
